requirements.txt
asgiref==3.2.10
Django==3.1
django-crispy-forms==1.9.2
django-multiselectfield==0.1.12
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
